+++ AdMob/src/main/java/com/mopub/mobileads/GooglePlayServicesBanner.java
@@ -178,10 +178,20 @@ public class GooglePlayServicesBanner extends BaseAd {
     @Override
     protected boolean checkAndInitializeSdk(@NonNull final Activity launcherActivity,
                                             @NonNull final AdData adData) {
+        Preconditions.checkNotNull(launcherActivity);
+        Preconditions.checkNotNull(adData);
+
         return false;
     }
 
     private class AdViewListener extends AdListener {
+        @Override
+        public void onAdImpression() {
+            if (mInteractionListener != null) {
+                mInteractionListener.onAdImpression();
+            }
+        }
+
         @Override
         public void onAdClosed() {
         }