+++ AdMob/src/main/java/com/mopub/nativeads/GooglePlayServicesAdRenderer.java
@@ -6,6 +6,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 import android.widget.ImageView;
+import android.widget.ImageView.ScaleType;
 import android.widget.TextView;
 
 import androidx.annotation.IdRes;
@@ -15,6 +16,7 @@ import androidx.annotation.Nullable;
 import com.google.android.gms.ads.nativead.AdChoicesView;
 import com.google.android.gms.ads.nativead.MediaView;
 import com.google.android.gms.ads.nativead.NativeAdView;
+import com.mopub.common.Preconditions;
 import com.mopub.common.logging.MoPubLog;
 import com.mopub.nativeads.GooglePlayServicesNative.GooglePlayServicesNativeAd;
 
@@ -91,6 +93,8 @@ public class GooglePlayServicesAdRenderer implements MoPubAdRenderer<GooglePlayS
     @NonNull
     @Override
     public View createAdView(@NonNull Context context, @Nullable ViewGroup parent) {
+        Preconditions.checkNotNull(context);
+
         View view = LayoutInflater.from(context).inflate(mViewBinder.layoutId, parent, false);
         // Create a frame layout and add the inflated view as a child. This will allow us to add
         // the Google native ad view into the view hierarchy at render time.
@@ -105,6 +109,9 @@ public class GooglePlayServicesAdRenderer implements MoPubAdRenderer<GooglePlayS
     @Override
     public void renderAdView(@NonNull View view,
                              @NonNull GooglePlayServicesNativeAd nativeAd) {
+        Preconditions.checkNotNull(view);
+        Preconditions.checkNotNull(nativeAd);
+
         GoogleStaticNativeViewHolder viewHolder = mViewHolderMap.get(view);
         if (viewHolder == null) {
             viewHolder = GoogleStaticNativeViewHolder.fromViewBinder(view, mViewBinder);
@@ -192,6 +199,7 @@ public class GooglePlayServicesAdRenderer implements MoPubAdRenderer<GooglePlayS
         nativeAdView.setBodyView(staticNativeViewHolder.mTextView);
         if (staticNativeViewHolder.mMediaView != null) {
             MediaView mediaview = new MediaView(nativeAdView.getContext());
+            mediaview.setImageScaleType(ScaleType.CENTER_CROP);
             staticNativeViewHolder.mMediaView.removeAllViews();
             staticNativeViewHolder.mMediaView.addView(mediaview);
             nativeAdView.setMediaView(mediaview);
@@ -226,6 +234,8 @@ public class GooglePlayServicesAdRenderer implements MoPubAdRenderer<GooglePlayS
 
     @Override
     public boolean supports(@NonNull BaseNativeAd nativeAd) {
+        Preconditions.checkNotNull(nativeAd);
+
         return nativeAd instanceof GooglePlayServicesNativeAd;
     }
 
@@ -262,6 +272,9 @@ public class GooglePlayServicesAdRenderer implements MoPubAdRenderer<GooglePlayS
         public static GoogleStaticNativeViewHolder fromViewBinder(@NonNull View view,
                                                                   @NonNull GooglePlayServicesViewBinder
                                                                           viewBinder) {
+            Preconditions.checkNotNull(view);
+            Preconditions.checkNotNull(viewBinder);
+
             final GoogleStaticNativeViewHolder viewHolder = new GoogleStaticNativeViewHolder();
             viewHolder.mMainView = view;
             try {