+++ AdMob/CHANGELOG.md
@@ -1,4 +1,20 @@
 ## Changelog
+  * 20.1.0.0
+   * This version of the adapters has been certified with AdMob 20.1.0 and MoPub 5.16.4.
+
+  * 20.0.0.2
+   * Remove mopub-sdk as a required dependency from POM file.
+
+  * 20.0.0.1
+    * Set MediaView's scale type to `CENTER_CROP` to fix native ad alignment issues.
+
+  * 20.0.0.0
+    * This version of the adapters has been certified with AdMob 20.0.0 and MoPub 5.16.4.
+
+  * 19.8.0.0
+    * This version of the adapters has been certified with AdMob 19.8.0 and MoPub 5.16.3.
+    * Align impression tracking for non-native ad formats based on Google's `onAdImpression()` callback.
+
   * 19.7.0.1
     * Service release. No code changes.