+++ AppLovin/CHANGELOG.md
@@ -1,4 +1,19 @@
 ## Changelog
+* 10.2.1.0
+    * This version of the adapters has been certified with AppLovin 10.2.1 and MoPub SDK 5.16.4.
+
+* 10.2.0.1
+    * Remove mopub-sdk as a required dependency from POM file.
+
+* 10.2.0.0
+    * This version of the adapters has been certified with AppLovin 10.2.0 and MoPub SDK 5.16.4.
+
+* 10.1.2.0
+    * This version of the adapters has been certified with AppLovin 10.1.2 and MoPub SDK 5.16.4.
+
+* 10.1.1.0
+    * This version of the adapters has been certified with AppLovin 10.1.1 and MoPub SDK 5.16.3.
+
 * 10.0.1.0
     * This version of the adapters has been certified with AppLovin 10.0.1 and MoPub SDK 5.16.0.