+++ Pangle/CHANGELOG.md
@@ -1,4 +1,16 @@
 ## Changelog
+  * 3.5.1.1.0
+     * This version of the adapters has been certified with Pangle 3.5.1.1 and MoPub SDK 5.17.0.
+
+  * 3.5.1.0.0
+     * This version of the adapters has been certified with Pangle 3.5.1.0 and MoPub SDK 5.16.4.
+     
+  * 3.5.0.5.1
+    * Remove mopub-sdk as a required dependency from POM file.
+
+  * 3.5.0.5.0
+     * This version of the adapters has been certified with Pangle 3.5.0.5 and MoPub SDK 5.16.4.
+
   * 3.4.1.1.1
      * Pass `AdCallSource` data for Pangle SDK to collect ad source information.